refactor: Resolve WPS214 split large classes
