+++ cli/core/products/services/related_components_base_service.py
@@ -1,7 +1,7 @@
 import logging
 from abc import ABC
 from collections.abc import Callable
-from typing import override
+from typing import Any, override
 
 from cli.core.errors import MPTAPIError
 from cli.core.models import DataCollectionModel
@@ -13,7 +13,78 @@ from cli.core.services.service_result import ServiceResult
 logger = logging.getLogger(__name__)
 
 
-class RelatedComponentsBaseService(RelatedBaseService, ABC):
+class RelatedComponentsActionMixin:
+    """Provide related-component action handlers."""
+
+    api: Any
+
+    def _action_create_item(self, data_model: DataModel):
+        """Creates the item in the API.
+
+        This method could be overridden by subclasses to customize the data model before
+        sending to API.
+
+        Args:
+            data_model: The data model to be created
+
+        """
+        new_data_model = self.api.post(json=data_model.to_json())
+        data_model.id = new_data_model["id"]  # type: ignore[attr-defined]
+
+    def _action_delete_item(self, _data_model: DataModel) -> None:
+        """Delete the item in the API.
+
+        This method could be overridden by subclasses to customize the data model before
+        sending to API.
+
+        Args:
+            _data_model: The data model to be deleted
+
+        """
+        logger.debug("Delete action is not supported yet")
+
+    def _action_update_item(self, data_model: DataModel) -> None:
+        """Update the item in the API.
+
+        This method could be overridden by subclasses to customize the data model before
+        sending to API.
+
+        Args:
+            data_model: The data model to be updated
+
+        """
+        self.api.update(data_model.id, data_model.to_json())  # type: ignore[attr-defined]
+
+    def _get_update_action_handler(self, model_action: DataActionEnum) -> Callable:
+        """Retrieve the appropriate action handler based on the action type.
+
+        This method could be overridden by subclasses to add specific actions in subclasses.
+
+        Args:
+            model_action: The action type to retrieve the handler for.
+
+        Returns:
+            Callable: The action handler for the given action type.
+
+        Raises:
+            ValueError: If the action type is not supported.
+
+        """
+        if model_action == DataActionEnum.CREATE:
+            return self._action_create_item
+
+        if model_action == DataActionEnum.DELETE:
+            # TODO: uncomment once the delete action is supported
+            # return self._action_delete_item   # noqa: ERA001
+            raise ValueError(f"Action type {model_action} is not supported")
+
+        if model_action == DataActionEnum.UPDATE:
+            return self._action_update_item
+
+        raise ValueError(f"Invalid action: {model_action}")
+
+
+class RelatedComponentsBaseService(RelatedComponentsActionMixin, RelatedBaseService, ABC):
     """Base service for managing related component operations."""
 
     @override
@@ -123,68 +194,3 @@ class RelatedComponentsBaseService(RelatedBaseService, ABC):
              DataModel: The data model to create
         """
         return data_model
-
-    def _action_create_item(self, data_model: DataModel):
-        """Creates the item in the API.
-
-        This method could be overridden by subclasses to customize the data model before
-        sending to API.
-
-        Args:
-            data_model: The data model to be created
-
-        """
-        new_data_model = self.api.post(json=data_model.to_json())
-        data_model.id = new_data_model["id"]  # type: ignore[attr-defined]
-
-    def _action_delete_item(self, data_model: DataModel) -> None:
-        """Delete the item in the API.
-
-        This method could be overridden by subclasses to customize the data model before
-        sending to API.
-
-        Args:
-            data_model: The data model to be deleted
-
-        """
-        logger.debug("Delete action is not supported yet")
-
-    def _action_update_item(self, data_model: DataModel) -> None:
-        """Update the item in the API.
-
-        This method could be overridden by subclasses to customize the data model before
-        sending to API.
-
-        Args:
-            data_model: The data model to be updated
-
-        """
-        self.api.update(data_model.id, data_model.to_json())  # type: ignore[attr-defined]
-
-    def _get_update_action_handler(self, model_action: DataActionEnum) -> Callable:
-        """Retrieve the appropriate action handler based onf the action type.
-
-        This method could be overridden by subclasses to add specific actions in subclasses.
-
-        Args:
-            model_action: The action type to retrieve the handler for.
-
-        Returns:
-            Callable: The action handler for the given action type.
-
-        Raises:
-            ValueError: If the action type is not supported.
-
-        """
-        if model_action == DataActionEnum.CREATE:
-            return self._action_create_item
-
-        if model_action == DataActionEnum.DELETE:
-            # TODO: uncomment once the delete action is supported
-            # return self._action_delete_item   # noqa: ERA001
-            raise ValueError(f"Action type {model_action} is not supported")
-
-        if model_action == DataActionEnum.UPDATE:
-            return self._action_update_item
-
-        raise ValueError(f"Invalid action: {model_action}")
