+++ pyproject.toml
@@ -121,8 +121,19 @@ select = ["AAA", "E999", "WPS"]
 show-source = true
 statistics = false
 per-file-ignores = [
-  "*.py:WPS210,WPS213,WPS214,WPS218,WPS221,WPS229,WPS231,WPS237,WPS335,WPS339,WPS342,WPS432",
-  "tests/**:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432"
+  "*.py:WPS210,WPS213,WPS221,WPS229,WPS231,WPS237,WPS335,WPS339,WPS342,WPS432",
+  "tests/**:WPS202,WPS204,WPS210,WPS213,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/handlers/test_excel_file_handler.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/price_lists/models/test_item.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/price_lists/models/test_price_list.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_item_group.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_items.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_parameter_group.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_parameters.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_product.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/models/test_template.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/core/products/test_app.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432",
+  "tests/cli/plugins/audit_plugin/test_audit_records.py:WPS202,WPS204,WPS210,WPS213,WPS218,WPS221,WPS335,WPS339,WPS342,WPS432"
 ]
 
 [tool.ruff]
